Safari-609.1.11.1/LayoutTests/fast/dom/rel-list-gc-expected.txt

Tests relList gc for anchor element
PASS element.relList.life is 42
PASS element.relList.life is 42
Tests relList gc for area element
PASS element.relList.life is 42
PASS element.relList.life is 42
Tests relList gc for link element
PASS element.relList.life is 42
PASS element.relList.life is 42
PASS successfullyParsed is true

TEST COMPLETE
